CoreModules/IAM/infrastructure/compliance/graphql/resolvers.py
"""
INNOVABIZ - Resolvedores GraphQL para Compliance IAM
==================================================================
Autor: Eduardo Jeremias
Data: 07/05/2025
Versão: 1.0
Descrição: Implementação dos resolvedores GraphQL para o serviço de 
           validação de compliance do módulo IAM, incluindo validação 
           HIPAA para o módulo Healthcare.
==================================================================
"""

import json
import uuid
import logging
from datetime import datetime, timedelta
from typing import Dict, List, Any, Optional, Union

import graphene
from graphene import Field, List as GList, String, ID, Int, Float, Boolean, Enum
from graphql import GraphQLError

from ..validator import (
    ComplianceFramework, 
    RegionCode, 
    ComplianceLevel,
    ComplianceValidatorFactory,
    MultiRegionComplianceValidator
)

# Configuração de logger
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("innovabiz.iam.compliance.graphql")

# ============================================================================
# Tipos Escalares
# ============================================================================

class JSONObject(graphene.Scalar):
    """Tipo escalar para representar objetos JSON genéricos"""
    
    @staticmethod
    def serialize(obj):
        return obj
    
    @staticmethod
    def parse_literal(node):
        return node.value
    
    @staticmethod
    def parse_value(value):
        return value

class DateTime(graphene.Scalar):
    """Tipo escalar para representar data e hora"""
    
    @staticmethod
    def serialize(dt):
        if isinstance(dt, datetime):
            return dt.isoformat()
        return dt
    
    @staticmethod
    def parse_literal(node):
        return datetime.fromisoformat(node.value)
    
    @staticmethod
    def parse_value(value):
        return datetime.fromisoformat(value)

# ============================================================================
# Enumerações
# ============================================================================

class ComplianceFrameworkEnum(graphene.Enum):
    """Enumeração de frameworks de compliance suportados"""
    GDPR = "gdpr"
    LGPD = "lgpd"
    HIPAA = "hipaa"
    PCI_DSS = "pci_dss"
    SOX = "sox"
    ISO_27001 = "iso_27001"
    NIST_800_53 = "nist_800_53"
    LAI = "lai"

class RegionCodeEnum(graphene.Enum):
    """Enumeração de regiões suportadas"""
    EU = "eu"
    PT = "pt"
    BR = "br"
    US = "us"
    AO = "ao"
    CD = "cd"
    GLOBAL = "global"

class ComplianceLevelEnum(graphene.Enum):
    """Enumeração de níveis de compliance"""
    COMPLIANT = "compliant"
    PARTIALLY_COMPLIANT = "partially_compliant"
    NON_COMPLIANT = "non_compliant"
    NOT_APPLICABLE = "not_applicable"

class ReportFormatEnum(graphene.Enum):
    """Enumeração de formatos de relatório"""
    HTML = "html"
    PDF = "pdf"
    JSON = "json"
    CSV = "csv"
    MARKDOWN = "markdown"

# ============================================================================
# Tipos de Objetos
# ============================================================================

class ComplianceFrameworkInfo(graphene.ObjectType):
    """Tipo para informações sobre framework de compliance"""
    code = String(required=True)
    name = String(required=True)
    name_pt = String(required=True)
    description = String(required=True)
    description_pt = String(required=True)
    region_scope = String()
    industry_scope = String()
    version = String()
    effective_date = DateTime()
    status = String(required=True)

class RegionInfo(graphene.ObjectType):
    """Tipo para informações sobre região"""
    code = String(required=True)
    name = String(required=True)
    name_pt = String(required=True)
    applicable_frameworks = GList(ComplianceFrameworkInfo, required=True)

class ComplianceRequirementType(graphene.ObjectType):
    """Tipo para requisito de compliance"""
    id = ID(required=True)
    code = String(required=True)
    title = String(required=True)
    description = String(required=True)
    category = String(required=True)
    severity = String(required=True)
    framework = Field(ComplianceFrameworkInfo, required=True)
    control_type = String()
    validation_logic = String()

class ValidationResult(graphene.ObjectType):
    """Tipo para resultado de validação de compliance"""
    success = Boolean(required=True)
    report_id = ID()
    timestamp = DateTime(required=True)
    issues = Int(required=True)
    compliant_requirements = Int(required=True)
    non_compliant_requirements = Int(required=True)
    partially_compliant_requirements = Int(required=True)
    not_applicable_requirements = Int(required=True)
    overall_compliance_score = Float()

class ReportDownload(graphene.ObjectType):
    """Tipo para URL de download de relatório"""
    format = Field(ReportFormatEnum, required=True)
    url = String(required=True)
    expires_at = DateTime(required=True)

class ReportGenerationResult(graphene.ObjectType):
    """Tipo para resultado de geração de relatório"""
    success = Boolean(required=True)
    report_id = ID(required=True)
    tenant = ID(required=True)
    timestamp = DateTime(required=True)
    formats = GList(ReportFormatEnum, required=True)
    download_urls = GList(ReportDownload, required=True)

class ComplianceReportSummary(graphene.ObjectType):
    """Tipo para resumo de relatório de compliance"""
    id = ID(required=True)
    tenant = ID(required=True)
    timestamp = DateTime(required=True)
    region = String()
    framework = String()
    overall_compliance_score = Float(required=True)
    status = String(required=True)
