CoreModules/IAM/requirements-observability.txt
# Requisitos para o módulo de observabilidade do IAM Audit Service
# INNOVABIZ - 2025

# Métricas e Exposição
prometheus-client>=0.16.0,<0.17.0
prometheus-fastapi-instrumentator>=6.0.0,<7.0.0

# FastAPI
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.22.0,<0.23.0

# Banco de Dados
psycopg[binary]>=3.1.10,<4.0.0
sqlalchemy>=2.0.20,<3.0.0
asyncpg>=0.27.0,<0.28.0

# Cache
redis>=4.6.0,<5.0.0
aioredis>=2.0.1,<3.0.0

# Rastreamento Distribuído
opentelemetry-api>=1.18.0,<1.20.0
opentelemetry-sdk>=1.18.0,<1.20.0
opentelemetry-exporter-otlp>=1.18.0,<1.20.0
opentelemetry-instrumentation-fastapi>=0.38b0,<0.40.0
opentelemetry-instrumentation-httpx>=0.38b0,<0.40.0

# Utilidades
pydantic>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
httpx>=0.24.1,<0.25.0